requirements.txt
python-dotenv==1.0.0
google-generativeai
mcp
google-genai
